Added `Configurator#addLoaderInclude` for more compatible configuring
The babel-loader was using an exclude Condition for node_modules.  This meant you never could include from node_modules, even as an exception.
To allow for this exception, preferred configuration is now to specify `src/**` as an include.
The helper method addLoaderInclude will see the glob, interpret it as relative, and automatically transform it to an absolute regex.